refactor: using one view model globaly
