+++ Task2/Task2/Task2/Views/EditMemo.swift
@@ -7,30 +7,18 @@
 
 import SwiftUI
 
-class EditMemoScreenViewModel: ObservableObject {
-    @Published var title: String = ""
-    @Published var content: String = ""
-    
-    func addMemo(to memoListViewModel: MemoListViewModel) {
-        memoListViewModel.addMemo(title: title, content: content)
-        title = ""
-        content = ""
-    }
-}
-
 struct EditMemoScreen: View {
-    @StateObject var viewModel: EditMemoScreenViewModel
-    @EnvironmentObject private var memoListViewModel: MemoListViewModel
+    @StateObject var memoListViewModel: MemoListViewModel
     @Environment(\.presentationMode) var presentationMode
     
     var body: some View {
         VStack(spacing: 20) {
             ZStack (alignment: .leading){
-                TextField("", text: $viewModel.title)
+                TextField("", text: $memoListViewModel.title)
                     .background(.white)
                     .textFieldStyle(.roundedBorder)
                 
-                if viewModel.title.isEmpty {
+                if memoListViewModel.title.isEmpty {
                     Text("제목을 작성해주세요")
                         .padding(.leading, 10)
                         .foregroundColor(TEXTCOLOR)
@@ -40,11 +28,11 @@ struct EditMemoScreen: View {
             
             ZStack(alignment: .topLeading) {
                 Color.gray.opacity(0.2)
-                TextEditor(text: $viewModel.content)
+                TextEditor(text: $memoListViewModel.content)
                     .background(Color.clear)
                     .scrollContentBackground(.hidden)
                     .padding()
-                if viewModel.content.isEmpty {
+                if memoListViewModel.content.isEmpty {
                     Text("내용을 작성해주세요")
                         .foregroundColor(TEXTCOLOR)
                         .padding(.all, 25)
@@ -58,7 +46,7 @@ struct EditMemoScreen: View {
         .toolbar {
             ToolbarItem(placement: .navigationBarTrailing) {
                 Button(action: {
-                    viewModel.addMemo(to: memoListViewModel)
+                    memoListViewModel.addMemo()
                     presentationMode.wrappedValue.dismiss()
                 }) {
                     Text("등록")
